fix get the version of matrixdb (#2)
It is true that `MatrixDB` has been officially named "YMatrix". However, for some historical reasons, while calling `pg_catalog.version()`, texts containing "MatrixDB" are still displaying for the client.
As a result, in the function we are to detect the version of `YMatrix` in order to decide which GUC names to use, we are still obliged to conform to the convention of legacy name of `MatrixDB`.