add contributing guide for developers (#3)
This PR adds a file named `CONTRIBUTING.md` to guide the developers through contributing to our project.

The content illustrates what they can do for the project, and the basic workflows to do it. It also includes some rules for commits and how to sign-off the work.

The rules are general for the majority of the open source projects based on Github.