fix 3 small issues in README (#8)
Fix several issues in the README:

For the sample config file, there is an inconsistency problem with the comment and the parameter.
The comment suggests that it will offer 3 queries, while the actual parameter only demonstrates one.
This PR dedicates to fix the comment, making it consistent with the parameter.
the comment for generator-disorder-ratio in the sample. Actually, the disordered timestamp will be 24 hours earlier.
Last but not least, fix the redundant digits of a figure for query latency.