exclude dump from query only mode (#13)
Previously, a "query-only mode"  is designed for `mxbench`. This happened when the "writer" or "generator" is "nil", implying that the user only wants `mxbench` to run queries. Thus, the step for creating table and loading data are skipped.

This is not supposed to be applied in the "dump" mode, in which the user would probably set "writer" to "nil" but anticipate generating data. So in this PR, we exclude "dump" from "query-only mode".